Separated creation of bootstrap null distributions and p-value calculation for bootstrap pvalue For bootstrap p-value calculation, added a fitting functor for the simple fit wrapper Single-ensemble simple fit wrapper can now compute covariance matrix from rawDataDistributions
fit_simple_AMA now allows for the user to provide a "multiplicity" file containing the weighting of the exact timeslices. This is to allow for random timeslice sampling for multiple exact timeslices to land on the same timeslice. If a multiplicity file is not provided and the number of exact timeslices in the data file is less than the amount specified by the user, random multiplicities will be assigned

fit_simple now produces a guess template when TEMPLATE is provided as the filename

Added some new data file formats and combinations supporting analysis of Grid_meas output using fit_simple_AMA / fit_simple